Reworked payment negotiation
With more stricter ordering and resending of messages, the protocol for
negotiating a new channel status can get simplified. As resending lost
messages means that a started exchange can not be aborted anymore, we
can already define the next revocation hash at the end of the current
exchange. With a known revocation hash, the initiator can directly send
signatures for the update.

Also, LNPaymentProcessor and LNPaymentLogic have been made less
stateful with relying on the database as single source of truth. This
means that much logic has been moved into the DBHandler.

+++ thunder-core/src/main/java/network/thunder/core/communication/layer/high/payments/messages/ChannelUpdate.java
@@ -12,7 +12,7 @@ public class ChannelUpdate implements Cloneable {
     public List<PaymentData> redeemedPayments = new ArrayList<>();
 
     public int feePerByte;
-    public long csvDelay;
+    public int csvDelay;
 
     public void applyConfiguration (LNConfiguration configuration) {
         this.feePerByte = configuration.DEFAULT_FEE_PER_BYTE;
